+++ db/migrate/20221220115303_add_source_site_relation_to_users.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddSourceSiteRelationToUsers < ActiveRecord::Migration[7.0]
+  def change
+    add_reference :folio_users, :source_site
+  end
+end