+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 2021-12-20
 ### Added
 - added `splittable_by_attribute` to atoms
+- added `source_site` relation to users
 
 ## 2021-12-13
 ### Added